move files to components, top scroll button bg correct
