--- app/src/app/portfolio/portfolio.module.css
@@ -1,3 +1,0 @@
-.projectButton:hover{
-  text-shadow: 1px 1px 1px rgba(6, 6, 6, 0.1);
-}
