+++ app/src/components/portfolio/ProjectRoot.tsx
@@ -1,6 +1,8 @@
 import React, { ReactNode } from "react";
 import { Box, Group, Image, Stack, Text, Title, rem } from "@mantine/core";
 
+import classes from "./portfolio.module.css";
+
 export interface Font {
   name: string;
   style: string;
@@ -50,10 +52,13 @@ const ProjectRoot: React.FC<ProjectRootProps> = ({
         bd="0.5px solid rgba(0, 0, 0, 0.2)"
         w="100%"
         h={400}
-        style={{ "border-radius": "16px", overflow: "clip" }}>
+        style={{ borderRadius: "16px", overflow: "clip" }}>
         {props.header}
       </Box>
-      <Group justify="space-between" align="start">
+      <Group
+        justify="space-between"
+        align="start"
+        className={classes.projectRootWrapGroup}>
         <Group gap={rem(32)} align="start" w="full">
           <Title order={3} w={{ base: "100%", md: rem(240) }}>
             &quot;{title}&quot;
@@ -63,7 +68,7 @@ const ProjectRoot: React.FC<ProjectRootProps> = ({
           <DescriptionText text={props.details} />
         </Group>
         <Stack
-          w={{ base: "100%", lg: rem(240) }}
+          w={{ base: "100%", md: rem(240) }}
           align="end"
           gap={0}
           mt={rem(-4)}>
