+++ app/src/components/portfolio/TheDrop.tsx
