+++ app/src/components/portfolio/portfolio.module.css
@@ -1,0 +1,13 @@
+.projectButton {
+  &:hover {
+    text-shadow: 1px 1px 1px rgba(6, 6, 6, 0.1);
+  }
+}
+
+.projectRootWrapGroup {
+  flex-wrap: wrap;
+
+  @media (min-width: $mantine-breakpoint-md) {
+    flex-wrap: nowrap;
+  }
+}
